crud status with animations and done update

+++ src/app/admin/status/status.component.ts
@@ -1,13 +1,38 @@
-import {Component, OnInit, Input, Inject} from '@angular/core';
+import {Component, OnInit, Input, Inject, ViewChild} from '@angular/core';
 import {ApiService} from '../../services/api.service';
 import {ISubscription} from 'rxjs/Subscription';
 import {DOCUMENT} from "@angular/common";
+import {animate, state, style, transition, trigger} from "@angular/animations";
 
 @Component({
     selector: 'app-status',
-    templateUrl: './status.component.html'
+    templateUrl: './status.component.html',
+    animations: [
+      trigger('list', [
+        state('in', style({
+          opacity: 1,
+          transform: 'translateX(0)'
+        })),
+        transition('void => *', [
+          style({
+            opacity: 0,
+            transform: 'translateX(-100px)'
+          }),
+          animate(300)
+        ]),
+        transition('* => void', [
+          animate(300, style({
+            opacity: 0,
+            transform: 'translateX(100px)'
+          }))
+        ]),
+      ])
+    ]
 })
 export class StatusComponent implements OnInit {
+
+
+  @ViewChild('inputCreate') inputCreate;
 
     statuses = [];
 
@@ -27,21 +52,24 @@ export class StatusComponent implements OnInit {
     ngOnInit() {
     }
 
-    delete(id: number) {
+    remove(i: number) {
+      const id = this.statuses[i].id;
+      if(confirm("Are you sure to delete: "+this.statuses[i].title)) {
         this.apiService.deleteStatus(id)
-            .subscribe(
+          .subscribe(
             (status) => {
-                console.log(status);
+              console.log(status);
             }
-        );
+          );
         this.statuses = this.statuses.filter( status => status.id !== id);
+      }
     }
 
     showUpdate(i) {
       this.statuses[i].edit = true;
     }
 
-    createStatus(title: string) {
+    create(title: string) {
         this.apiService.createStatus(title).subscribe(
             (role) => {
                 console.log(role);
@@ -50,14 +78,18 @@ export class StatusComponent implements OnInit {
         );
     }
 
+    cancel(i: number) {
+      const id = this.statuses[i].id;
+      (<HTMLInputElement>this.doc.getElementById('input-title-'+id)).value = this.statuses[i].title;
+      this.statuses[i].edit = false;
+    }
 
     update(i: number) {
-
       const id = this.statuses[i].id;
       const title = (<HTMLInputElement>this.doc.getElementById('input-title-'+id)).value;
 
       this.statuses[i].title = '';
-        this.apiService.updateStatus( status.title).subscribe(
+        this.apiService.updateStatus(id, title).subscribe(
             (role) => {
                 this.statuses[i].title = title;
                 this.statuses[i].edit = false;
